+++ av1an/chunk/Chunk.py
@@ -38,7 +38,6 @@ class Chunk:
         self.frames: int = frames
         self.output_ext: str = output_ext
         self.per_shot_target_quality_cq = None
-        self.per_frame_target_quality_q_list = None
 
     def to_dict(self) -> Dict[str, Any]:
         """